// src/store/main/resource/center/types.ts
/*
 * @Author: korealu
 * @Date: 2022-02-16 16:53:07
 * @LastEditors: Please set LastEditors
 * @LastEditTime: 2022-04-13 10:02:13
 * @Description: file content
 * @FilePath: /pofi-admin/src/store/main/device/imei/types.ts
 */
export interface IResourceCenterType {
  centersList: any[]
  centersCount: number
  selectData: any
}